Refactor: sgasser/ember-cli-materialize/17
