+++ addon/components/materialize-button-submit.js
@@ -1,0 +1,8 @@
+import materializeButton from './materialize-button';
+
+export default materializeButton.extend({
+  layoutName: 'components/materialize-button',
+  tagName: 'button',
+  attributeBindings: ['type'],
+  type: 'submit'
+});
