+++ addon/components/materialize-card-action.js
@@ -1,0 +1,6 @@
+import Ember from 'ember';
+
+export default Ember.Component.extend({
+  tagName: 'div',
+  classNames: ['card-action']
+});
