+++ addon/components/materialize-card-panel.js
@@ -1,0 +1,10 @@
+import Ember from 'ember';
+import layout from '../templates/components/materialize-card-panel';
+
+export default Ember.Component.extend({
+  layout: layout,
+  tagName: 'div',
+  classNames: ['card-panel'],
+  classNameBinding: 'class'
+});
+
