+++ addon/components/materialize-card-reveal.js
@@ -1,0 +1,10 @@
+import Ember from 'ember';
+import layout from '../templates/components/materialize-card-reveal';
+
+export default Ember.Component.extend({
+  layout: layout,
+  tagName: 'div',
+  classNames: ['card-reveal'],
+  classNameBinding: 'class',
+  activatorBinding: 'parentView.activator'
+});
