+++ addon/components/materialize-collapsible.js
@@ -1,0 +1,8 @@
+import Ember from 'ember';
+import layout from '../templates/components/materialize-collapsible';
+
+export default Ember.Component.extend({
+  layout: layout,
+  tagName: 'li',
+  classNameBindings: 'class'
+});
