+++ addon/templates/components/materialize-card-content.hbs
