+++ addon/templates/components/materialize-card-panel.hbs
