+++ addon/templates/components/materialize-card-reveal.hbs
