+++ addon/templates/components/materialize-collapsible-card.hbs
