+++ addon/templates/components/materialize-collapsible.hbs
