+++ addon/templates/components/materialize-navbar.hbs
