+++ package.json
@@ -16,6 +16,9 @@
   },
   "author": "Stefan Gasser",
   "license": "MIT",
+  "dependencies": {
+    "ember-cli-htmlbars": "0.7.4"
+  },
   "devDependencies": {
     "broccoli-asset-rev": "^2.0.0",
     "ember-cli": "^0.2.0",
@@ -23,7 +26,6 @@
     "ember-cli-babel": "^4.0.0",
     "ember-cli-content-security-policy": "0.3.0",
     "ember-cli-dependency-checker": "0.0.8",
-    "ember-cli-htmlbars": "0.7.4",
     "ember-cli-ic-ajax": "0.1.1",
     "ember-cli-inject-live-reload": "^1.3.0",
     "ember-cli-qunit": "0.3.9",
